Extract Kore.Strategies.Rule from Kore.Strategies.Goal
